문자열 처리와 연산부 클래스 분리
