+++ src/main/kotlin/calculator/Calculator.kt
@@ -17,7 +17,7 @@ class Calculator {
         return a / b
     }
 
-    private fun calculate(a: Int, op: String, b: Int): Int {
+    fun calculate(a: Int, op: String, b: Int): Int {
         return when (op) {
             "+" -> add(a, b)
             "-" -> subtract(a, b)
@@ -25,18 +25,5 @@ class Calculator {
             "/" -> divide(a, b)
             else -> throw IllegalArgumentException("Invalid operator : $op")
         }
-    }
-
-    fun run(formula: String?): Int {
-        val trimmed = formula?.trim()
-        if (trimmed.isNullOrEmpty()) throw IllegalArgumentException("Formula should not be null or empty")
-
-        val tokens = trimmed.split(" ")
-        var result = tokens.first().toInt()
-        for (i in 1 until tokens.size step 2) {
-            result = calculate(a = result, op = tokens[i], b = tokens[i + 1].toInt())
-        }
-
-        return result
     }
 }
